spine/requirements.txt
pytest>=7.0.0
click>=8.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
